Refactor appointments container into calendar and form components
